requirements.txt
livekit-agents>=0.5.0
livekit>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
elevenlabs>=0.2.0
silero-vad>=4.0.0
